+++ changelog.md
@@ -19,6 +19,7 @@
 - IMPORTANT!!! Change all API Calls URL's.
 - Add Docker Network (`docker network create`) support. The configuration variable ist "DOCKER_CNI".
 - Seperate K3S Server and Agent Memory and CPU resource definition.
+- Change K3S to use docker engine
 
 ## v0.2.0